# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "0.1.0",
        "title": "Diary API",
        "description": "Open Childcare Platform",
        "contact": {
            "name": "Mash Software",
            "url": "https://www.mashsoftware.com",
            "email": "hello@mashsoftware.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:5000/v1",
            "description": "Local Development"
        },
        {
            "url": "https://mash-diary-api-staging.herokuapp.com/v1",
            "description": "Staging"
        },
        {
            "url": "https://mash-diary-api.herokuapp.com/v1",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/children": {
            "get": {
                "summary": "Retrieves a list of children",
                "operationId": "get_children",
                "tags": [
                    "Children"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Users ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paged array of children",
                        "headers": {
                            "x-next": {
                                "description": "A link to the next page of responses",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChildResponse"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates a new child",
                "operationId": "create_child",
                "tags": [
                    "Children"
                ],
                "requestBody": {
                    "description": "New child data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChildRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Newly created child",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChildResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/children/{id}": {
            "get": {
                "summary": "Retrieves a specific child",
                "operationId": "get_child",
                "tags": [
                    "Children"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id of the child to retrieve",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChildResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Updates a specific child",
                "operationId": "update_child",
                "tags": [
                    "Children"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id of the child to update",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChildResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletes a specific child",
                "operationId": "delete_child",
                "tags": [
                    "Children"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id of the child to delete",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expected response to a valid request"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChildRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "date_of_birth"
                ],
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date"
                    },
                    "users": {
                        "type": "array",
                        "minItems": 1,
                        "uniqueItems": true,
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                }
            },
            "ChildResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date"
                    },
                    "users": {
                        "type": "array",
                        "minItems": 1,
                        "uniqueItems": true,
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                }
            },
...
            "Error": {
                "required": [
                    "message"
                ],
                "properties": {
                    "description": {
                        "type": "string"
                    }
                }
            }
        }
    }
}